+++ lib/src/directus_api_manager_base.dart
@@ -745,7 +745,7 @@ class DirectusApiManager implements IDirectusApiManager {
       String? title,
       String? contentType,
       String? folder,
-      String storage = "local",
+      String? storage,
       Map<String, dynamic>? additionalFields}) {
     return _sendRequest(
         canSaveResponseToCache: false,